requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==7.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
gunicorn==21.2.0
